requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.2.1               # via kombu
asn1crypto==0.22.0        # via cryptography
bcrypt==3.1.3             # via paramiko
beautifulsoup4==4.6.0     # via wagtail
billiard==3.5.0.3         # via celery
boto==2.39.0
celery==4.0.2
cffi==1.10.0              # via bcrypt, cryptography, pynacl
contextlib2==0.5.5        # via raven
cryptography==2.0.2       # via paramiko
decorator==4.1.2          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid, social-auth-core
dj-database-url==0.3.0
dj-static==0.0.6
django-discover-runner==1.0  # via django-role-permissions
django-modelcluster==3.1  # via wagtail
django-redis==4.7.0
django-role-permissions==1.2
django-server-status==0.4.0
django-storages-redux==1.3.2
django-taggit==0.22.1     # via wagtail
django-treebeard==4.1.2   # via wagtail
django-webpack-loader==0.4.1
django==1.10.5
djangorestframework==3.5.2
edx-api-client==0.4.0
edx-opaque-keys==0.4
elasticsearch-dsl==2.1.0
elasticsearch==2.3.0
factory-boy==2.8.1
faker==0.7.18
gnupg==2.2.0
html5lib==0.999999999
idna==2.5                 # via cryptography
ipython-genutils==0.2.0   # via traitlets
ipython==6.1.0
jedi==0.10.2              # via ipython
jsonfield==1.0.3
kombu==4.1.0              # via celery, django-server-status
newrelic==2.88.1.73
oauthlib==2.0.2           # via requests-oauthlib, social-auth-core
open-discussions-client==0.2.0
paramiko==2.2.1           # via pysftp
pbr==3.1.1                # via stevedore
pexpect==4.2.1            # via ipython
phonenumbers==8.3.2
pickleshare==0.7.4        # via ipython
pillow==3.4.2
prompt-toolkit==1.0.15    # via ipython
psutil==5.2.2             # via gnupg
psycopg2==2.6
ptyprocess==0.5.2         # via pexpect
pyasn1==0.3.1             # via paramiko
pycountry==16.11.27.1
pycparser==2.18           # via cffi
pygments==2.2.0           # via ipython
pyjwt==1.5.2              # via open-discussions-client, social-auth-core
pymongo==3.4.0            # via edx-opaque-keys
pynacl==1.1.2             # via paramiko
pysftp==0.2.9
python-dateutil==2.5.3
python3-openid==3.1.0     # via social-auth-core
pytz==2017.2              # via celery, django-modelcluster
raven==6.1.0
redis==2.10.5
requests-oauthlib==0.8.0  # via social-auth-core
requests==2.11.1
robohash==1.0
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via bcrypt, cryptography, django-role-permissions, django-server-status, edx-api-client, edx-opaque-keys, elasticsearch-dsl, faker, html5lib, prompt-toolkit, pynacl, python-dateutil, social-auth-app-django, social-auth-core, stevedore, traitlets
social-auth-app-django==1.1.0
social-auth-core==1.4.0   # via social-auth-app-django
static3==0.5.1
stevedore==1.25.0         # via edx-opaque-keys
tornado==4.5.1            # via robohash
traitlets==4.3.2          # via ipython
unidecode==0.4.21         # via wagtail
urllib3==1.22             # via elasticsearch
uwsgi==2.0.15
vine==1.1.4               # via amqp
wagtail==1.10.1
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
willow==0.4               # via wagtail
